+++ service/firebase/users.go
@@ -3,9 +3,11 @@ package firebase
 import (
 	"context"
 	"fmt"
+	"os"
 	"time"
 
 	"bitbucket.org/andyfusniakteam/ecom-api-go/model/postgres"
+	"cloud.google.com/go/pubsub"
 	"firebase.google.com/go/auth"
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
@@ -147,6 +149,20 @@ func (s *Service) CreateUser(ctx context.Context, role, email, password, firstna
 		Created:     c.Created,
 		Modified:    c.Modified,
 	}
+
+	publishResult := s.eventsTopic.Publish(ctx, &pubsub.Message{
+		Attributes: map[string]string{
+			"event": "user.created",
+		},
+		Data: []byte("test"),
+	})
+	serverID, err := publishResult.Get(ctx)
+	if err != nil {
+		fmt.Fprintf(os.Stderr, "%+v", err)
+	}
+	fmt.Printf("serverID=%s\n", serverID)
+	contextLogger.Debugf("published a message to pubsub")
+
 	return &ac, nil
 }